pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xgboost-mono"
version = "0.1.0"
description = "XGBoost monotonic constraints demonstration"
requires-python = ">=3.12"
dependencies = [
    "kaggle>=1.6.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",
    "python-dotenv>=1.0.0",
    "xgboost>=2.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pyright>=1.1.0",
    "pre-commit>=3.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RET",    # flake8-return
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented out code)
    "PL",     # pylint
    "PERF",   # perflint
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
    "N803",   # argument name should be lowercase (allow X_train, X_test for ML convention)
    "N806",   # variable in function should be lowercase (allow X_train, X_test for ML convention)
    "PLR0915", # too many statements (grid_search is a long orchestration function)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]  # unused imports and import order in __init__.py

[tool.pyright]
include = ["src", "scripts"]
exclude = [".venv", "**/__pycache__"]
typeCheckingMode = "standard"
pythonVersion = "3.12"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
